+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.8] 2021-07-12
+
+- Add Extensions in Error of `InputValueValidator`. [#564](https://github.com/async-graphql/async-graphql/pull/564)
+
+- Fix SDL print is not stable. [#547](https://github.com/async-graphql/async-graphql/issues/547)
+
 ## [2.9.7] 2021-07-04
 
 - Add support for generic `ComplexObject`. [#562](https://github.com/async-graphql/async-graphql/pull/562)